# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sporteam API",
    "version": "2.0.0",
    "description": "API RESTful para la gestión de jugadores, entrenadores y clubs de fútbol amateur",
    "contact": {
      "name": "Roman Dono Perez"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desarrollo"
    }
  ],
  "components": {
    "securitySchemes": {
      "TokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "token",
        "description": "Token JWT de autenticación"
      }
    },
    "schemas": {
...
      "Entrenador": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nombre": {
            "type": "string"
          },
          "apellidos": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string"
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "estadoDeportivo": {
            "type": "string",
            "nullable": true
          },
          "clubId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "entrenador": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "nombreDeportivo": {
                "type": "string",
                "nullable": true
              },
              "telefono": {
                "type": "string",
                "nullable": true
              },
              "entrenadorPorteros": {
                "type": "boolean"
              },
              "titulacion": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "EntrenadorInput": {
        "type": "object",
        "required": [
          "nombre",
          "apellidos",
          "email",
          "password"
        ],
        "properties": {
          "nombre": {
            "type": "string"
          },
          "apellidos": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 6
          },
          "role": {
            "type": "string",
            "enum": [
              "ENTRENADOR_ROLE"
            ]
          },
          "estadoDeportivo": {
            "type": "string"
          },
          "zona": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "club": {
            "type": "string"
          },
          "nombreDeportivo": {
            "type": "string"
          },
          "entrenadorPorteros": {
            "type": "boolean"
          },
          "titulacion": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "telefono": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/entrenadores": {
      "get": {
        "tags": [
          "Entrenadores"
        ],
        "summary": "Listar entrenadores con paginación",
        "security": [
          {
            "TokenAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "desde",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "in": "query",
            "name": "limite",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de entrenadores",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "total": {
                      "type": "integer"
                    },
                    "entrenadores": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Entrenador"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error al recuperar entrenadores"
          }
        }
      }
    },
    "/api/entrenadores/{termino}": {
      "get": {
        "tags": [
          "Entrenadores"
        ],
        "summary": "Buscar entrenadores por término",
        "security": [
          {
            "TokenAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "termino",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resultados de búsqueda",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "resultados": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Entrenador"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/entrenador/{id}": {
      "get": {
        "tags": [
          "Entrenadores"
        ],
        "summary": "Obtener entrenador por ID de usuario",
        "security": [
          {
            "TokenAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Entrenador encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "entrenador": {
                      "$ref": "#/components/schemas/Entrenador"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Entrenador no encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Entrenadores"
        ],
        "summary": "Actualizar un entrenador",
        "security": [
          {
            "TokenAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombreDeportivo": {
                    "type": "string"
                  },
                  "entrenadorPorteros": {
                    "type": "boolean"
                  },
                  "titulacion": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "telefono": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Entrenador actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "entrenador": {
                      "$ref": "#/components/schemas/Entrenador"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Entrenador no encontrado"
          }
        }
      }
    },
    "/api/entrenador": {
      "post": {
        "tags": [
          "Entrenadores"
        ],
        "summary": "Crear un nuevo entrenador (usuario + perfil)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntrenadorInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Entrenador creado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "entrenador": {
                      "$ref": "#/components/schemas/Entrenador"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error de validación"
          }
        }
      }
    },
    "/api/entrenadores/zona/{idZona}": {
      "get": {
        "tags": [
          "Entrenadores"
        ],
        "summary": "Listar entrenadores por zona",
        "security": [
          {
            "TokenAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "idZona",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "desde",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "in": "query",
            "name": "limite",
            "schema": {
              "type": "integer",
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Entrenadores filtrados por zona",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "total": {
                      "type": "integer"
                    },
                    "entrenadores": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Entrenador"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}